requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
httpx==0.27.2
python-dateutil==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
itsdangerous==2.2.0
python-dotenv==1.0.1
cryptography==42.0.8
